update asset fields
