+++ stac_pydantic/item.py
@@ -9,12 +9,12 @@ from pydantic.fields import FieldInfo
 from .api.extensions.context import ContextExtension
 from .api.extensions.paging import PaginationLink
 from .extensions import Extensions
-from .shared import Asset, BBox, ExtensionTypes, Link
+from .shared import Asset, BBox, ExtensionTypes, Link, StacCommonMetadata
 from .utils import decompose_model
 from .version import STAC_VERSION
 
 
-class ItemProperties(BaseModel):
+class ItemProperties(StacCommonMetadata):
     """
     https://github.com/radiantearth/stac-spec/blob/v0.9.0/item-spec/item-spec.md#properties-object
     """
@@ -22,15 +22,6 @@ class ItemProperties(BaseModel):
     datetime: Union[str, dt] = Field(..., alias="datetime")
     created: Optional[Union[str, dt]] = Field(None, alias="datetime")
     updated: Optional[Union[str, dt]] = Field(None, alias="datetime")
-    # stac common metadata (https://github.com/radiantearth/stac-spec/blob/v0.9.0/item-spec/common-metadata.md)
-    title: Optional[str] = Field(None, alias="title")
-    description: Optional[str] = Field(None, alias="description")
-    start_datetime: Optional[Union[str, dt]] = Field(None, alias="start_datetime")
-    end_datetime: Optional[Union[str, dt]] = Field(None, alias="end_datetime")
-    platform: Optional[str] = Field(None, alias="platform")
-    instruments: Optional[List[str]] = Field(None, alias="instruments")
-    constellation: Optional[str] = Field(None, alias="constellation")
-    mission: Optional[str] = Field(None, alias="mission")
 
     class Config:
         extra = "allow"
